Additions to part 2, real world experiments and glaciers as water resources.
First draft of part 2, contains a mix of the glacier_water_resources notebook and the new hydro outputs described in the hydro_outputs notebook. Explains how to do a commitment run, "run_with_climate_data", and how to open and plot some of the outputs. Myabe we want to add something more about temperature bias runs to further explain the role of equilibirum and water availablity.